Title: 평균점수 구하기, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240723/평균점수 구하기/print-average.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main() {
-    int score[8];
+    double score[8];
     double sum = 0;
     double avg = 0;
